autorag-system/requirements.txt
# NLP & Embedding
sentence-transformers==2.2.2
transformers==4.41.1
torch>=1.13.0

# Türkçe NLP
stanza==1.8.2

# PDF, DOCX, TXT ayrıştırma
pypdf==3.17.4
python-docx==1.1.0
unstructured==0.12.6

# Vektör veritabanı (Qdrant)
qdrant-client==1.8.0

# Metin işleme & benzerlik
scikit-learn==1.4.2
numpy==1.26.4
tqdm==4.66.2

# HTTP istekleri (Ollama için)
requests==2.31.0
